simplify audio example

+++ _s0il_/programs/httpd.c
@@ -1,3 +1,8 @@
+// TODO : basic-auth
+//        adapt CSS for running on smart-phones
+//        decode markdown
+
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -67,9 +72,17 @@ static filemapping filemappings[]={
   {NULL, NULL},
 };
 
+// XXX : this is hairy - perhaps having a baseroot where it
+//       is permitted is less hairy?
+bool allow_filesystem_modification = true;
+
+// see comment above
+bool allow_run                     = true;
+
 static char httpd_buf2[1024 * 32];
 static char httpd_buf[2048];
 
+const char *html_doctype="<!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Strict//EN' 'http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd'>\n";
 static char *httpd_css = 
 "body { background:black; color:white;}\n"
 "a { color:white;}\n"
@@ -275,7 +288,7 @@ static void httpd_browse_handler (HttpdRequest *req)
       }
 
   req->mime_type = "text/html";
-  OUTS( "<!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Strict//EN' 'http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd'>\n");
+  OUTS(html_doctype);
   OUTF( "<html><head><title>%s</title>\n<style type='text/css'>\n%s</style>\n", req->path, httpd_css);
     OUTS( "<link rel='stylesheet' href='/sd/webassets/codemirror.css'/>\n"
           "<link rel='stylesheet' href='/sd/webassets/codemirror-cobalt.css'/>\n"
@@ -388,19 +401,15 @@ static void httpd_browse_handler (HttpdRequest *req)
      }
     else if (!strcmp (mime_type, "inode/directory"))
     {
-
-
-  append_dir_listing(req, di, "/_");
-  dir_listing_destroy (di);
-
+        append_dir_listing(req, di, "/_");
+        dir_listing_destroy (di);
     }
     else
-     {
+    {
         
-     }
+    }
 
     OUTS( "</div>\n"); // view
-
     OUTS( "</form>\n");
   }
 
@@ -419,41 +428,21 @@ static void httpd_browse_handler (HttpdRequest *req)
 
 static void httpd_dir_handler (HttpdRequest *req)
 {
-  dir_listing_t *di = dir_listing_read(req, req->path);
-
   req->mime_type = "text/html";
-
-  OUTS( "<!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Strict//EN' 'http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd'>\n");
+  const char *html_doctype="<!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Strict//EN' 'http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd'>\n";
+  OUTS(html_doctype);
   OUTF( "<html><head><title>%s</title>\n<style type='text/css'>\n%s</style>\n", req->path, httpd_css);
 
   OUTS( "</head><body>");
 
+  {
+  dir_listing_t *di = dir_listing_read(req, req->path);
   append_dir_listing (req, di, "");
-#if 0
-  OUTS( "<ul class='dir_listing'>");
-
-  for (int i = 0; i < di->count; i++)
-   {
-     dir_entry_t *de = &di->entries[i];
-     const char *mime_type = ui_get_mime_type (NULL, de->path);
-
-     if (!strcmp (mime_type, "inode/directory"))
-     {
-       OUTF( "<li><a href='%s/'><span class='name'>%s</span></a>/</li>", de->path, de->name, mime_type);
-     }
-     else
-     OUTF( "<li><a href='%s'><span class='name'>%s</span></a>  <span class='mime_type'>%s</span></li>", de->path, de->name, mime_type);
-   }
-  OUTS( "</ul>");
-#endif
   dir_listing_destroy (di);
+  }
 
   OUTS( "</body></html>");
-
-  //free (path);
 }
-
-
 
 
 typedef enum action_t {
@@ -550,9 +539,8 @@ httpd_serve_file (HttpdRequest *req, const char *path)
            if (to_read > (int)sizeof(buf))to_read = sizeof(buf);
            read = fread (buf, 1, to_read, file);
            if (read>0) {
-             //printf ("%i of %i to %p\n", read_total, content_length, req->f);
-           fwrite (buf, 1, read, req->f);
-           read_total += read;
+             fwrite (buf, 1, read, req->f);
+             read_total += read;
            }
         } while ((read > 0) && (read_total < content_length));
         fclose (file);
@@ -589,6 +577,9 @@ static void httpd_post_handler (HttpdRequest *req)
      memcpy (var, p, len);\
      var[len]=0;\
   }
+
+  printf ("{%s}\n", req->body->str);
+
   get_arg(post_param, "param=")
   get_arg(post_origname, "origname=")
   get_arg(post_path, "path=")
@@ -627,11 +618,14 @@ static void httpd_post_handler (HttpdRequest *req)
       break;
 
     case action_rmdir:
-        snprintf (temp, sizeof(temp), "%s%s", post_path, post_param);
-        rmdir(temp);
+        if (allow_filesystem_modification)
+        {
+          snprintf (temp, sizeof(temp), "%s%s", post_path, post_param);
+          rmdir(temp);
+        }
       /* FALLTHROUGH */
     case action_remove:
-      if (action == action_remove)
+      if (allow_filesystem_modification && action == action_remove)
       {
         snprintf (temp, sizeof(temp), "%s%s", post_path, post_param);
         unlink(temp);
@@ -642,12 +636,12 @@ static void httpd_post_handler (HttpdRequest *req)
     case action_run:
       req->status = 301;
 
-      if ((action == action_save || action==action_run)
-           && post_content && strlen(post_content)>1)
+      if (allow_filesystem_modification && ((action == action_save || action==action_run)
+           && post_content && strlen(post_content)>1))
       {
       snprintf (temp, sizeof(temp), "Location: /_%s%s", post_path, post_param);       
 
-      if (post_content){
+      if (allow_filesystem_modification && post_content){
       FILE *f = fopen (temp+12, "w");
       if(f)
       {
@@ -667,10 +661,9 @@ static void httpd_post_handler (HttpdRequest *req)
       }
       }
 
-      if (action == action_run)
+      if (allow_run && action == action_run)
       {
         char *commandline = strdup (temp+12);
-        //printf ("%i\n", system(commandline));
         printf ("%i\n", runs(commandline));
         free (commandline);
       }
@@ -687,12 +680,15 @@ static void httpd_post_handler (HttpdRequest *req)
       req->extra_headers = temp;
       break;
     case action_rename:
+      if (allow_filesystem_modification)
+      {
       req->status = 301;
       char buf[512];
       snprintf (buf, sizeof(buf), "%s%s", post_path, post_origname);
       snprintf (temp, sizeof(temp), "Location: /_%s%s", post_path, post_param);
       rename(buf, temp + 12);
       req->extra_headers = temp;
+      }
       break;
     case action_reload:
       req->status = 301;
@@ -700,7 +696,7 @@ static void httpd_post_handler (HttpdRequest *req)
       req->extra_headers = temp;
       break;
     case action_mkfile:
-      {
+      if (allow_filesystem_modification){
       req->status = 301;
       snprintf (temp, sizeof(temp), "Location: /_%s%s", post_path, post_param);
       char *content = "";
@@ -715,7 +711,7 @@ static void httpd_post_handler (HttpdRequest *req)
       break;
 
     case action_mkdir:
-      {
+      if(allow_filesystem_modification){
         req->status = 301;
         snprintf (temp, sizeof(temp), "Location: /_%s%s/", post_path, post_param);
         mkdir (temp+12, 0777);//"w");
@@ -732,6 +728,24 @@ static void httpd_post_handler (HttpdRequest *req)
   ctx_string_set (req->body, "");
 
   return;
+}
+
+static void
+httpd_request_handler_put (HttpdRequest *req)
+{
+  FILE *file = fopen (req->path, "w");
+  if (file)
+  {
+    req->status = 200;
+    fwrite(req->body->str, 1, req->body->length, file);
+    fclose(file);
+    ctx_string_set (req->body, "file uploaded");
+  }
+  else
+  {
+    req->status = 500;
+    ctx_string_set (req->body, "failed");
+  }
 }
 
 static void
@@ -788,8 +802,6 @@ httpd_request_handler (HttpdRequest *req)
       OUTF( "404 no spoon %i %s", rno++, req->path);
    }
 }
-
-
 
 static void request_init (HttpdRequest *req, FILE *f)
 {
@@ -881,7 +893,7 @@ int _httpd_start_int (int port,
 #else
       void *foo,
 #endif
-                                    void *user_data)
+      void *user_data)
 {
   int    sock;
   struct sockaddr_in sin;
@@ -915,9 +927,6 @@ int _httpd_start_int (int port,
   httpd_stop = 0;
   while (!httpd_stop) /* should have some terminating condition.. */
     {
-#if 0 //USE_FORK
-      int pid;
-#endif
       HttpdRequest req;
       char *tmp;
       int content_length = -1;
@@ -932,20 +941,9 @@ int _httpd_start_int (int port,
 #else
         vTaskDelay(1);
 #endif
-        //printf("!\n");
         continue;
       }
 
-#if 0 //USE_FORK
-      pid = fork ();
-      if (pid < 0)
-        {
-          fprintf (stderr, "error on fork\n");
-          continue;
-        }
-      if (pid == 0) /* this is the client process */
-        {
-#endif
           f = fdopen (s, "a+");
           if (!fgets (httpd_buf, sizeof (httpd_buf), f))
             continue;
@@ -970,7 +968,6 @@ int _httpd_start_int (int port,
                   content_length = atoi (sep);
                 }
               }
-
             if (!strncmp (httpd_buf2, "\r", strlen ("\r")))
               // end of headers
               break;
@@ -985,10 +982,9 @@ int _httpd_start_int (int port,
                 int read_attempt = sizeof(httpd_buf2);
                 if (content_length - read_total < read_attempt)
                   read_attempt = content_length - read_total;
-              read = fread (httpd_buf2, 1, read_attempt, f);
-              ctx_string_append_data (req.body, httpd_buf2, read);
-              read_total += read;
-              //printf ("read: %i %i of %i\n", read, read_total, content_length);
+                read = fread (httpd_buf2, 1, read_attempt, f);
+                ctx_string_append_data (req.body, httpd_buf2, read);
+                read_total += read;
               } while (read>0 && read_total < content_length);
 
               if ((int)read_total != content_length)
@@ -1003,33 +999,26 @@ int _httpd_start_int (int port,
               }
             }
           fseek (f, 0, SEEK_CUR); 
-          if (!strcasecmp (req.method, "GET") ||
-              !strcasecmp (req.method, "POST"))
+          if (allow_filesystem_modification && !strcasecmp (req.method, "PUT"))
+          {
+            httpd_request_handler_put (&req);
+          }
+          else if (!strcasecmp (req.method, "GET") ||
+                   !strcasecmp (req.method, "POST"))
             httpd_request_handler (&req);
           else
             {
               req.status = 501;
               req.status_string = "Not supported";
-              ctx_string_append_printf (req.body, "only POST or GET");
+              ctx_string_set (req.body, "unsupported http method");
             }
           request_finish (&req);
           fclose (f);
-#if 0//USE_FORK
-          exit (0);
-        }
-      else
-        {
-          close (s);
-        }
-#else
       close (s);
-#endif
     }
   close (sock);
   return 0;
 }
-
-#include <signal.h>
 
 int main (int argc, char **argv)
 {
